pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finance"
version = "0.1.0"
description = "Personal Finance Tracking System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.13",

    # Web Framework
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "jinja2>=3.1",
    "python-multipart>=0.0.6",

    # PDF Processing
    "pikepdf>=8.0",
    "pdfplumber>=0.10",
    "tabula-py>=2.9",

    # Data Processing
    "python-dateutil>=2.8",
    "rapidfuzz>=3.6",

    # Configuration
    "pydantic-settings>=2.1",

    # Utilities
    "click>=8.1",
    "python-dotenv>=1.0",
    "httpx>=0.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
]

[project.scripts]
finance = "finance.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/finance"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
